[admin-tool] Add support to dump UPDATE records and RMD records in Kafka Topic Dumper (#831)
This PR adds new mode to print data record (PUT/DELETE/UPDATE) as well as its corresponding RMD record to console, on top of its existing functionality of logging Kafka message metadata to console. It is very useful for debugging feature correctness.

In the future, we can consider adding a key filter to log specific key only, to better focus on certain key.