[da-vinci][server] Fixed a couple of issues with lossy rewind check
We discovered 3 issues with lossy rewind check:
1. The value comparison is not correct since the value decoded from KME is a ByteBuffer,
   and the backed array can contain non-value bytes.
2. For the batch push, the storage engine is not usable until receiving EOP, where Venice Server
   would ingest all the generated sst files into storage engine, so before that, storage engine
   lookup will always return null.
3. The lossy rewind function is currently being invoked after persisting current record into
   local storage engine, which means the check will always return true.

This PR fixes 1 & 2, and 3 requires more refactoing, and I will try to fix it in a different PR.

For the batch push to a hybrid store, and if split-brain issue happens inside the batch portion,
the push job will fail because the hybrid version topic will enable log compaction later, and the
log compaction can generate out-of-order messages and DIV sequence will be wrong as well.
Handling this situation will be very tricky, so we would like to work on a formal solution
to address split-brain issue for such use case.