[common] update http client socket timeout configs (#854)
With low QPS, the clients close the connection between client and venice-server if its idle for a minute leading to new SSL Handshakes and latency spikes. Revert the httpclient5 configs added in PR 829 and change ConnectionConfig.setSocketTimeout() from 1 minute to 0(indefinite) such that client won't initiate closing the connection if it is inactive for any period of time.