[controller] Fixed NPE in StoreInfo::fromStore
When this function receives a null Store parameter, it will return null itself.

There are four calls to this function, and they have all been inspected to
ensure that they deal properly with nulls.

In the case of VHA::migrateStore, there are some usability improvements so that
the VeniceNoStoreException includes the cluster the store is in, if it is in a
different one than the supplied source cluster. There is a new short-circuiting
logic as well in case the store is already in the desired destination cluster.