[router] Fixed NPE in ReadRequestThrottler
There could be a race condition where a listener of the ReadRequestThrottler got
called while the constructor was not fully done executing, leading to a NPE for
accessing a field that was not yet initialized. The fix is to register listeners
at the very end.

Also eliminated the RoutingDataRepository constructor param and class property
since it was unused.