[da-vinci] Added deprecated flag to DaVinciConfig#NonLocalAccessPolicy
This PR is the first step to deprecate NonLocalAccessPolicy#QUERY_VENICE.
This feature will be completely removed in a later release.
Here are the reasons this feature doesn't work well:
 1. Remote Venice query is much slower than DaVinci local lookup.
 2. Typically, Venice Backend doesn't provision enough capacity for DaVinci use cases since we normally don't know
    the qps of DaVinci apps and for most of the scenarios, the DaVinci qps can be very high, and we don't want to
    pre-allocate a lot of backend resources for this feature since the remote resources won't be leveraged most of
    the time.

In the future, if there are DaVinci use cases, who would like to leverage such capability, we will need to evaluate
it properly to make sure such requirement is reasonable or not.
From a different perspective, if we can improve DaVinci bootstrapping performance significantly by the blob file transfer, such
requirement won't exist IMO.