[controller] Parent controller should not update the current version (#862)
Currently, there are some codepaths that lead to parent controller updating the current version. Another unexpected code path that gets triggered in parent controller is RealTimeTopicSwitcher#transmitVersionSwapMessage. Since this piece wasn't written with the intention of getting executed in parent controller, it causes failures in executing tasks from the admin channel. This causes the Admin channel to get stuck for this store.

This commit adds checks that prevent the current version in parent controller to be updated.