[vpj] Add Spark as a compute engine for DataWriter job (#851)
This PR adds Spark as an alternate compute engine to run the Data Writer task. Currently, it does not handle KIF and the code is written in a way that it will fall back to the MapReduce implementation in case KIF is used. This will be addressed in a follow up PR.

To use the new Spark based flow, set the following VPJ config:
"data.writer.compute.job.class = com.linkedin.venice.hadoop.spark.datawriter.jobs.DataWriterSparkJob"

VPJ has the following new configs:
1. "venice.spark.cluster": This will configure the spark.master config.
    * Refer to https://spark.apache.org/docs/latest/submitting-applications.html#master-urls for the supported values
2. "venice.spark.session.conf.*"
    * Configs with this prefix will be set when building the spark session
    * These will get applied to all Spark jobs that get triggered as a part of VPJ
    * It can be used to configure arbitrary cluster properties
3. "spark.data.writer.conf.*"
    * Configs with this prefix will be set when building the data writer spark job and passed as job properties
    * These will only get applied on the DataWriter Spark jobs
    * It is useful when there are custom input formats which need additional configs to be able to read the data