[common] Fix an infinite recursion issue in VeniceWriter
Today, VeniceWriter (VW) has a bug to run into an infinite recursion when it sends
a message in the venice writer and eventually it throws StackOverflowError.
For a particular partition, the conditions to trigger the issue are:

1. VW has a cached Segment
2. The segment is neither started nor ended
3. The elapsed time for the segment is greater than MAX_ELAPSED_TIME_FOR_SEGMENT_IN_MS

To fix it, we break the infinite loop by always recording the new start time.