[vpj] Fix deterministic GUID generation for VPJ (#891)
Recently, when abstracting compute engines, the logic to compute a deterministic GUID was changed to be less reliant on the underlying compute engine. However, this change had a bug where the uniqueness of the GUID was tied to the uniqueness of the "jobId" used to trigger the VenicePushJob. This commit generates a unique GUID per VPJ job.