[server][da-vinci] Remove noise when clearing offsets without metadata partition
When a version is retired, both Helix resource and ingestion task will be deleted/killed;
when Helix resource is deleted, OFFLINE->DROPPED transition message will be sent to server;
and when ingestion task is killed, a kill-job message will be sent to servers; these two
deletions are happening independently; if transitions are done before kill-job, and if
ingestion task is trying to flush RESET_OFFSET actions, servers will fail to find metadata
partition.

If metadata partition doesn't exist when trying to clear offsets, there is nothing left to
clear and thus no point throwing exception for it.