[vpj] Allow enabling compression metric collection without using mapper to build dictionary (#886)
Currently, "use.mapper.to.build.dictionary" is auto enabled when "compression.metric.collection.enabled" is enabled. This commit separates the two configs for migration purposes and "use.mapper.to.build.dictionary" will not be auto-enabled when "compression.metric.collection.enabled" is enabled and it now needs to be explicitly enabled. Eventually, "use.mapper.to.build.dictionary" will be removed along with the "ValidateSchemaAndBuildDict" job.

Operationally, we have noticed that often, the VPJ jobs wait in the queue to acquire resources from Yarn despite the job only requesting a container for one map task and no reduce tasks. When abstracting the compute engines used by VPJ, we decided to not abstract the "ValidateSchemaAndBuildDict" job. We decided this because the initial reasons to add a dedicated job are no longer valid:
1. Zstd dictionary creation would sometimes crash the JVM (https://github.com/luben/zstd-jni/issues/253)
    * This issue has now been fixed in the library; and
    *  We also have guardrails for this in our code
2. Ability to build larger dictionaries: 
    * We believe the VPJ executors can have enough memory to allow them to build good enough dictionaries
    * At LinkedIn, we have found that dictionaries of size <150 KB and using 200 MB of samples offer great compression ratios. (Upto 40X in some cases). And for this, our executors only need 256 MB of memory.