[controller] Added status update timestamps to JobStatusQueryResponse.
A Lazy<JobStatusQueryResponse> is added as input to all hooks which return the
StoreVersionLifecycleEventOutcome. This is to facilitate the implementation of
the hook to inject a pause between each region swap.

Added @Deprecated to some of the values in the ExecutionStatus enum.

Added a blurb in the StoreLifecycleHooks's JavaDoc to specify that invocations
for the hook of a given store are guaranteed to be serialized.

Specified that incremental push is out of scope of this proposal.