[server] Server read quota usage ratio metric improvement
1. Previously the quota ratio is only updated whenever the token bucket is refilled.
This is problematic for stores that have a low KPS relative to their allocated quota since
it won't be refilled very often. As a result we could see the old usage ratio for extended
periods of time if the token bucket is not refilled frequently.

2. To improve this we will perform the calculation whenever the metric is read by using the
requested token since last refill and elasped time since last refill to compute the average KPS
and divide by refill KPS.