[davinci] Don't throw DIV MISSING data fault for new segment after EOP
Today, for hybrid stores after EOP is received, DIV creates a new segment and
throws data missing fault when:

1. EOP is received
2. no previous segment is found for the input guid and partition in the record
3. sequence number in the record is > 0

Actually, this does not seem to be right simply because that by looking at the data
that DIV has at the moment, it's too early to make the conclusion that missing data
scenario happened (e.g. this is the very first message that DIV receives from
the guid for that partition).

In this change, DIV accepts that this is a new message from the guid for the partition
and not throws any fault. (also a bit of code refactoring in related function)