[changelog] Make changelog consumer creation parallel in VeniceChangelogConsumerClientFactory (#907)
In user onboarding we noticed that multiple consumer will spent extra time on creation. During log investigation, we found that this factory method is the root cause.
The synchronized keyword makes consumer creation sequential, including Apache consumer creation and StorageService restoration. Each of them could take 1-2s so in total it took ~11 in practice to complete.
This change should make it truly parallel in creation and save ~10s during bootstrap.