[Controller] Implement dual read for Push Status System Store
With changes in this PR, controller will try to compose a version level
key and use it to read from push status system store when checking batch
push status on DaVinci side:
{
  keyStrings: {10} // Version 10
  messageType: 1 // 1: Full Push
}

If controller gets null response for the new key, it will assume that
DaVinci is not upgraded yet, and fall back to partition level keys:
{
  keyStrings: {10, 0} // Version 10, partition 0
  messageType: 1 // 1: Full Push
}

Roll out steps in order:
1. After the controller changes are rolled out to production and stabilize,
   deprecate all previous backend releases.
2. Release DaVinci changes to start producing a version level batch updates;
   stabilize the client release and deprecate all previous client releases.
3. Clean up tech debts in controller - remove all partition level keys related
   logic.

Other changes:
Fixed a bug in the partition level DaVinci status computation - previously
when a small amount of instances go offline, the push can still complete
and version swap would happen. This PR fixed the bug and ensure that
as long as there is one offline instance, complete will not be reported.