[changelog] Add RocksDB block cache metric to changeLogClientConfig; Expose RocksDBMemoryStats to changelogConsumer (#920)
After customer adopts multiple changelog consumer, we encounter OOM and application killed by consumer. Based on current config setup, we suspect it is caused by multiple RocksDB instance opening 16GB block cache and eventually it use more memory than it should. Since we don't expose RocksDBMemoryStats today, we cannot make sure this is 100% the root cause, but from the behavior change from 1 consumer to 8 consumer with the same dataset, it is likely the cause.

Personally I think it will be great if multiple consumer can share same RocksDB instance without performance impact during bootstrap and consumption. Based on current implementation it would be hard to do so, so we will put this PR as an incremental improvement.

This PR opens up the config for user to set per consumer cache size limit. We will verify from the new metrics in the new client release.