[server] Calculate upstream offset from TopicSwitch if needed when transitioning into LEADER (#926)
PR #908 fixed the bug that follower may calculate and store wrong rewind offset if some RT messages are gone because of message retention.
However, the fix is not complete. It missed the case during leader transition: The to-be-leader will store TS (if it sees one), it will update the leader topic, but it will not calculate the rewind and persist it after the PR change. The result is that: when it becomes leader after 5 min inactive time, it will start from -1, as no upstream offset has been computed.

This PR proposes to fix the case by adding the same offset calculation logic to API startConsumingAsLeader.

It happens to improve the resiliency of AA subscription as well by reusing the same code. As in the past, the to-be-leader will not re-try calculating offset for unreachable broker via repair task.