[Controller][all] Do not truncate Kafka topic immediately for fatal data validation error before EOP
When the ingestion of a new store version is failed, today, we truncate the Kafka topic of the
store version by updating its retention time to a small value (15 seconds), specified by
DEPRECATED_TOPIC_RETENTION_MS.

This is fine for regular failures but for fatal data validation errors, where it indicates critical issues
happened during the batch push period (before EOP), truncating the topic too early can prevent us from
finding the root cause, simply because the kafka data is gone.

This change adjusts the retention time (to 5 days) when fatal data validation errors is identified, so
that we can have enough time to investigate.