[dvc] Do not throw DIV segment MISSING when a host regaining leadership after EOP
DIV dumps its segments information into offset record periodically or before server shut down
and loads the contents back during Helix state transition (OFFLINE->STANDBY) when the host is
transitioned to follower and, when it becomes the leader, copied into the leader DIV.

For hybrid store partitions (after EOP), if a host went through a LEADER -> (OFFLINE ->) FOLLOWER
-> LEADER transition, e.g. because of host shutdown, its DIV dumps and reloads segments as above
described. However, today, if another host became the new leader (when it was offline or follower),
did its leadershipduty by consuming messages from RT topic, producing to local VT etc, then when
the previous host became the leader for the 2nd time, its DIV, after loading segment info from
checkpoint, only contains the stale segment info in terms of the segment number for the RT topic,
thus when it starts to consume, it will throw DIV MISSING error, which in most cases a false alarm.

This false alarm can happen very offten as long as for a hybrid store partition, a host shuts down,
rt topic keeps getting new data and consumed by new leader, old host starts again and regains the
leadership for the same rt topic.

This rb introduces a flag to indicate if a segment is constructed from a check point. When a host regains
its leadership and reconsumes RT, DIV tolerates the segment gap for very first message it meets and
creates a new unregistered segment (based on the contents fo the impending message) to replace it.