[fast-client] Long tail retry budget for fast client (#930)
* [fast-client] Long tail retry budget for fast client

1. Introduced RetryManager to measure and calculate retry budget based on a percentage of
requests.

2. The RetryManager is only applied to long tail retries and currently hard coded to be 10%
of the total original request rate. This means if there were any exception that's not a 429
the retry will be triggered without going through the RetryManager. If the retry budget is
exhausted then the retry task will do nothing and the request will either complete
eventually (original future) or time out.