[changelog] Fix after-image change capture client dropping record issue (#954)
VeniceChangeLogConsumerImpl tries to perform topic switch after receiving EOP. It is a NoOp for after-image CDC, as it should always consume from VT. However, it should not drop unprocessed records in the same polled batch which contains EOP control message.

This PR fixes the issue by adding a simple check to see whether it actually switch the topic. If TS does not actually happen, the consumer should finish processing all the data records it polled.