[changelog] Make bootstrapping consumer based on afterimage consumer
This is kind of a bad refactor, but theres only one user of this for now.  This changes the base change capture consumer from using the before image consumer to using the after image consumer.  The reason why we change the base class is because the afterimage consumer does do a little extra work for detecting version swap.  In light of this, extending is possibly a poor way to implement this class?  We might prefer to wrap somehows, but we implemented it this way so we could intersect some calls.... oh well