[server] - Fix consumer leaks in TM that occur when the thread owning consumer is interrupted (#967)
Currently, if a thread owning a consumer is interrupted, the consumer is not 
returned to the consumer pool. This occurs because the blocking calls throw an 
interrupted exception when invoked by a thread with the interrupt status set. 
In this case, it is the `LBQ::put` call in `TMDF::releaseConsumer`.

As a result, we eventually run out of consumers to execute metadata requests. 
This situation typically arises when non-blocking metrics collection threads 
cancel active executions to terminate slow queries.