[vpj] Fix building or fetching dictionary for repush jobs (#971)
Recently, we refactored some code around compression metric collection and using mapper to build dictionary. This refactoring had a bug where the dictionary building for repush jobs was skipped and if VPJ sends control messages directly, then the push would fail, as the "START_OF_PUSH" would not have any dictionary. This change fixes the issue and reenables a dictionary to be built for KIF repush as well.