[test] Fix for cross-test interrupt contamination
There is an issue either in Gradle or TestNG where a test which
sets the interrupt flag on the current thread can cause a later
test to fail, due to checking for the interrupt flag (e.g. in
sleep and syncrhonization-related primitives).

This commit fixes this by clearing the interrupt flag via the
VeniceTestListener::onTestStart listener. Also registered this
listener to unit tests (it was previously used only for integ
tests), and added a minimal unit test which deterministically
reproduces this issue (in the absence of the listener).

Miscellaneous:

- Added calls to super in all functions of VeniceTestListener,
  as recommended in the parent class' JavaDoc.