[server][dvc] Adopt standard logging format for topic-partition/replica ID to simplify log extraction - PART 1 (#989)
Currently, extracting logs for a specific replica is challenging due to varying log formats  
chosen by developers. For instance, consider the complexity of this command:  
grep -e " 205 " -e ": 205)" -e "=205," -e "=205 " -e ": 205 " -e "_205" -e "-205-" -e " 205;"  
-e ": 205," venice-server-war.log.2023-11-2* | grep -v "205 ms" | grep -v "205 second"  
Even with such a command, we might not retrieve all logs for the replica. The aim of this commit  
is to adopt a standardized logging format for topic-partition/replica ID, to make log extraction  
easier and more straightforward.