[router] Delete dictionary for retired versions instead of existing versions (#1010)
There is a bug where future for dictionary downloads of existing versions would be removed, but instead, the intention is to delete for retired versions. This PR fixed the bug.

A major change:
Previously it's the ZK callback thread that sends the requests to retrieve dictionary; now it's the router dictionary processing threads that send the requests and wait for response. This is the only way to make testing feasible and meaningful.