[all] Make getValueSchemaId APIs match the canonical schemas (#1018)
Previously, "getValueSchemaId" APIs in "SchemaReader", "SchemaData" and "ReadOnlySchemaRepository" match the schemas exactly (with the exception of "HelixReadWriteSchemaRepository"). These APIs are generally used to get a schema id to write along with the data. In such cases, we only care about canonical schemas.

The exception is where we want to register a schema with a doc change. In our codebase, this schema check is not handled via "getValueSchemaId" APIs and happens in the "VeniceHelixAdmin#addValueSchema" code paths which handle this case explicitly.

This commit makes "getValueSchemaId" use canonical schema match in all places.