[vpj] Unify shared temp directories for VPJ
Currently, VPJ uses temp directories for at least two features:
1. To store the output from ValidateSchemaAndBuildDict mapper job to pass the dictionary to the driver
2. To store schemas for TTL repush

In the future, we might add more such cases where data needs to be stored in a
temp directory. For operational reasons, it is desirable to have a temp directory that is shared by all
VPJ jobs, and inside this, we can create other feature-specific shared directories that are also shared
by all VPJ jobs. These shared directories will have 777 permissions so any user can write to it. If
features have private data that need restricted permissions, the feature implementation can create files
or subdirectories inside the feature directories and apply the restricted permissions to those.

After this commit, the the temp directory will be:
.
|____<hadoop.tmp.dir> (Specified by env, or default /tmp)
| |____venice-push-job (777 permissions) - shared by all VPJ
| | |____<job.execution.id>_<unique-suffix> (700 permissions) - shared by all features in this execution
| | | |____veniceMapperOutput (700 permissions)
| | | |____rmd_schemas (700 permissions)
| | | |____value_schemas (700 permissions)
| | | |____...<features_added_in the future> (700 permissions)