[vpj] Remove extra call to PubSub to fetch partition count for data writer job (#1030)
Remove extra call to PubSub to fetch partition count for data writer job. There was a job failure in production where the VeniceWriter tried to query PubSub to fetch the partition count during the PartitionWriter phase. Since the partition count is always available before the DataWriter job starts, we can specify it explicitly and reduce one call to the PubSub system.

Also, improve docs and logging for shared temp directory for VPJ