[all][test] Adds a new checksum type INCDIGEST for data integrity validation
This PR adds a new checksum type INCDIGEST for Venice's DIV. INCDIGEST adopts the
ideal of incremental hash from AdHash and leverages the CRC32 as the underlying
hasing algorithm.

CRC32 is commonly used in networks and storage devices for detecting accidental
changes. The purpose of this algorithm is not to protect against intentional
changes (cryptographic hashing), but rather to catch accidents like network errors
and disk write errors, etc. The emphasis of CRC32 is more on speed than on security.

For testing, this PR extends all the existing checksum and DIV tests to include the
new INCDIGEST type.