[all] Remove Amplification Factor implementation in ingestion logics (#1035)
This PR does the following changes:
(1) Remove all amplification factor ingestion logics from VPJ, SystemProducer and Ingestion code, as well as serving code. The ingestion lifecycle should not play with AmpFactor and SubPartition concept anymore. (There is still a SubPartitionStatus Enum, which is now aiming at partition level, we can rename it in the future).
(2) Remove the ability to update amplification factor in admin-tool
(3) Does NOT delete Store/Version/PartitionConfig's amplification factor field and most of the controller setter/getter logic related to amp factor. This will be further cleaned up in the future, as it is not critical part to clean up ingestion architecture and unblock team's work.
(4) Remove ingestion isolation's logic to transfer LEADER state, as it is only in DVC now, and DVC will only do FOLLOWER mode.