[controller] Set ACTIVE_ACTIVE as default DRP when converting store to AA
For AA-enabled stores, it is crucial to consider lag from all fabrics to avoid
premature version swaps in some colos. To achieve this, we need to use the
ACTIVE_ACTIVE data replication policy. Currently, ACTIVE_ACTIVE is set only when
converting batch stores to hybrid AA-enabled stores. This PR updates the default
policy to ACTIVE_ACTIVE when an existing batch store is made AA-enabled.